% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cap_major_topics}
\alias{cap_major_topics}
\title{Major topic codes from the Comparative Agendas Project}
\format{An object of class \code{data.frame} with 21 rows and 2 columns.}
\source{
See for more: https://www.comparativeagendas.net/datasets_codebooks
}
\usage{
data(cap_major_topics)
}
\description{
Major topic codes from the Comparative Agendas Project
}
\keyword{datasets}
