% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cap_subtopics}
\alias{cap_subtopics}
\title{Subtopic codes from the Comparative Agendas Project}
\format{An object of class \code{data.frame} with 213 rows and 3 columns.}
\source{
See for more: https://www.comparativeagendas.net/datasets_codebooks
}
\usage{
data(cap_subtopics)
}
\description{
Subtopic codes from the Comparative Agendas Project
}
\keyword{datasets}
