% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_house_NOMINATE.R
\name{get_house_NOMINATE}
\alias{get_house_NOMINATE}
\title{Get DW-NOMINATE Scores for members of the U.S. House}
\usage{
get_house_NOMINATE(congress = "ALL")
}
\arguments{
\item{congress}{Number for the congress you're interested in. Defaults to "ALL" to get the most recent scores for every congress.}
}
\value{
a data frame containing the relevant legislator names and DW-NOMINATE scores, as well as otherinformation made available by VoteView
}
\description{
Returns a datset of NOMINATE scores for House members of the the specified congress,
available otherwise at VoteView.com.
}
\examples{

\dontrun{

 ## get the scores for the 116th (2019-2021) congress
 house <- get_house_NOMINATE(congress=116)

 ## returned as a data frame
 house

}

}
\seealso{
https://voteview.com/data
}
