% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senate_nominate.R
\name{get_senate_nominate}
\alias{get_senate_nominate}
\title{Get DW-nominate Scores for members of the U.S. Senate}
\usage{
get_senate_nominate(congress = "ALL")
}
\arguments{
\item{congress}{Number for the congress you're interested in. Defaults to "ALL" to get the most recent scores for every congress.}
}
\value{
a data frame containing the relevant legislator names and DW-nominate scores, as well as otherinformation made available by VoteView
}
\description{
Returns a dataset of nominate scores for Senate members of the the specified congress,
available otherwise at VoteView.com.
}
\examples{

\dontrun{

 ## get the scores for the 116th (2019-2021) congress
 senate <- get_senate_nominate(congress=116)

 ## returned as a data frame
 senate

}

}
\seealso{
https://voteview.com/data
}
