% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_results}
\alias{house_results}
\title{US House election results from 1976 to 2018}
\format{An object of class \code{data.frame} with 9994 rows and 7 columns.}
\source{
See for more: https://electionlab.mit.edu and official sites
}
\usage{
data(house_results)
}
\description{
US House election results from 1976 to 2018
}
\keyword{datasets}
