% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pres_results}
\alias{pres_results}
\title{State-by-state US presidential election results from 1976 to 2016}
\format{An object of class \code{data.frame} with 561 rows and 6 columns.}
\source{
See for more: https://electionlab.mit.edu and official sites
}
\usage{
data(pres_results)
}
\description{
State-by-state US presidential election results from 1976 to 2016
}
\keyword{datasets}
