% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pres_results_by_cd}
\alias{pres_results_by_cd}
\title{Presidential election by congressional district}
\format{An object of class \code{data.frame} with 4201 rows and 7 columns.}
\source{
See for more: https://www.thealmanacofamericanpolitics.com
}
\usage{
data(pres_results_by_cd)
}
\description{
Presidential election by congressional district
}
\keyword{datasets}
