% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trump_approval_polls_538.R
\name{trump_approval_polls_538}
\alias{trump_approval_polls_538}
\title{Download FiveThirtyEight's dataset of Donald Trump's job approval polling}
\usage{
trump_approval_polls_538(subgroup = "All polls")
}
\arguments{
\item{subgroup}{Filters the dataset to the polled population, either 'Adults', 'All polls', or 'Voters'. Defaults to 'All polls'}
}
\value{
a data.frame of individual polls aggregated by FiveThirtyEight
}
\description{
Returns a data.frame of individual polls aggregated by FiveThirtyEight
}
\examples{

\donttest{

 # get all polls
 polls <- trump_approval_polls_538(subgroup='Adults')

 # returned as a data.frame
 polls

}

}
\seealso{
Their dataset at https://projects.fivethirtyeight.com/trump-approval-data/approval_polllist.csv and interactive at https://projects.fivethirtyeight.com/trump-approval-ratings/
}
