% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_pres_polls_history}
\alias{us_pres_polls_history}
\title{Every presidential general election poll from 1980 through 2016}
\format{An object of class \code{data.frame} with 3586 rows and 10 columns.}
\usage{
data(us_pres_polls_history)
}
\description{
Every presidential general election poll from 1980 through 2016
}
\keyword{datasets}
