% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_116}
\alias{house_116}
\title{DW-NOMINATE scores for the 116th House}
\format{An object of class \code{data.frame} with 434 rows and 22 columns.}
\usage{
data(house_116)
}
\description{
Data for:
   - house_116: the result of calling \code{`get_house_nominate(congress=116)`}
}
\keyword{datasets}
