% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senate_116}
\alias{senate_116}
\title{DW-NOMINATE scores for the Senators of the 116th Congress}
\format{An object of class \code{data.frame} with 100 rows and 22 columns.}
\usage{
data(senate_116)
}
\description{
Data for:
   - senate_116: the result of calling \code{`get_senate_nominate(congress=116)`}
}
\keyword{datasets}
