% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_pres_polls_history}
\alias{us_pres_polls_history}
\title{Pre-loaded data for basic analysis and tests}
\format{An object of class \code{data.frame} with 3586 rows and 10 columns.}
\usage{
data(senate_116)
}
\description{
Data for:
   - us_pres_poll_history: presidential election polls from 1980 through 2016
}
\keyword{datasets}
